+++ undine/server/driver/base_driver.py
@@ -1,7 +1,11 @@
-from undine.utils.exception import VirtualMethodException
+from undine.database.rabbitmq import RabbitMQConnector
+from undine.utils.exception import UndineException, VirtualMethodException
 from undine.utils.system import print_console_header
+from undine.utils.system import System
 
 import undine.utils.logging as logging
+
+import json
 
 
 class BaseDriver:
@@ -57,3 +61,77 @@ class BaseDriver:
 
     def is_ready(self):
         raise VirtualMethodException(self.__class__, '_wait_others')
+
+
+class BaseNetworkDriver(BaseDriver):
+    #
+    # Constructor & Destructor
+    #
+    def __init__(self, task_queue, config, config_dir):
+        BaseDriver.__init__(self, config, config_dir)
+
+        if task_queue is None:
+            raise UndineException('Missing RabbitMQ option field (task_queue)')
+
+        self._queue = RabbitMQConnector(task_queue)
+        self._host = System.host_info()
+
+        self._logged_in()
+
+    def __del__(self):
+        self._logged_out()
+
+    #
+    # Private method
+    #
+    def _make_params(self, tid):
+        return {'tid': tid, 'host': self._host.name, 'ip': self._host.ipv4}
+
+    @property
+    def host(self):
+        return self._host
+
+    #
+    # Protected inherited interface
+    #
+    def _task(self, _tid):
+        raise VirtualMethodException(self.__class__, '_task')
+
+    def _preempt(self, _info):
+        raise VirtualMethodException(self.__class__, '_preempt')
+
+    def _done(self, _info, _content, _report):
+        raise VirtualMethodException(self.__class__, '_done')
+
+    def _cancel(self, _info):
+        raise VirtualMethodException(self.__class__, '_cancel')
+
+    def _fail(self, _info, _message):
+        raise VirtualMethodException(self.__class__, '_fail')
+
+    def _logged_in(self):
+        raise VirtualMethodException(self.__class__, '_logged_in')
+
+    def _logged_out(self):
+        raise VirtualMethodException(self.__class__, '_logged_out')
+
+    #
+    # Public interface
+    #
+    def fetch(self):
+        return self._task(json.loads(self._queue.consume())['tid'])
+
+    def preempt(self, tid):
+        return self._preempt(self._make_params(tid))
+
+    def done(self, tid, content, report):
+        return self._done(self._make_params(tid), content, report)
+
+    def cancel(self, tid):
+        return self._cancel(self._make_params(tid))
+
+    def fail(self, tid, message):
+        return self._fail(self._make_params(tid), message)
+
+    def is_ready(self):
+        return True